YT_stream_record_DC_2_Notion/requirements.txt
discord.py
requests
tqdm
asyncio